# backend/src/main/java/com/eaglebank/backend/dto/UpdateUserRequest.java
package com.eaglebank.backend.dto;

import jakarta.validation.Valid;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.Pattern;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateUserRequest {
    // All fields are optional for PATCH; when provided they will be validated and applied
    private String name;

    @Email(message = "Invalid email format")
    private String email;

    @Pattern(regexp = "^\\+[1-9]\\d{1,14}$", message = "Invalid phone number format")
    private String phoneNumber;

    @Valid
    private AddressDto address; // If provided, must satisfy AddressDto validations

    // Allow password change; when provided it will be encoded
    private String password;
}